+++ chess.html
@@ -54,6 +54,6 @@
 	</div>
 </div>
 
-<script type="module" src="chess.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
