--- selectPieceState.js
@@ -1,17 +1,0 @@
-export const selectPieceState = {
-    isClicked: false,
-    selectedSquare: null,
-    selectedSquareId: null,
-    destinationSquare: null,
-    clickOnPieceToReset: [],
-
-    selectedPiece: null,
-    selectedPieceArray: null,
-    selectedPieceIndex: null,
-    pieceType: null,
-    valueInSquare: null,
-    pieceColor: null,
-
-    x_squareCoordinate: null,
-    y_squareCoordinate: null,
-}
